@@ -15,22 +15,22 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.7.1
+version: 1.8.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  sqflite: ^1.3.0
-  path_provider: ^1.6.24
-  shared_preferences: ^0.5.12+4
-  provider: ^4.3.2+2
-  intl: ^0.16.1
-  flutter_staggered_animations: ^0.1.3
+  sqflite:
+  path_provider:
+  intl:
+  flutter_staggered_animations:
   pie_chart: ^5.0.0
-  url_launcher: ^5.7.10
+  url_launcher:
+  jiffy:
+  easy_dynamic_theme:
 
 
   # The following adds the Cupertino Icons font to your application.
