@@ -3,7 +3,7 @@ import 'package:flutter/services.dart';
 import 'package:moodiefschmtz/pages/home.dart';
 
 class App extends StatefulWidget {
-  const App({Key key}) : super(key: key);
+  const App({Key? key}) : super(key: key);
 
   @override
   _AppState createState() => _AppState();
@@ -19,9 +19,14 @@ class _AppState extends State<App> {
   @override
   Widget build(BuildContext context) {
 
-    final Color bottomOverlayColor = Theme.of(context).bottomAppBarColor;
-    final Color topOverlayColor = Theme.of(context).appBarTheme.backgroundColor;
-    final Brightness iconBrightness = Theme.of(context).primaryColorBrightness;
+    final Color bottomOverlayColor =
+    Theme.of(context).bottomNavigationBarTheme.backgroundColor!;
+    final Color topOverlayColor =
+    Theme.of(context).appBarTheme.backgroundColor!;
+    final Brightness iconBrightness =
+    Theme.of(context).brightness == Brightness.light
+        ? Brightness.dark
+        : Brightness.light;
 
     return AnnotatedRegion<SystemUiOverlayStyle>(
         value: SystemUiOverlayStyle(
