@@ -1,89 +1,120 @@
 import 'package:flutter/material.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
-//CLARO
 ThemeData light = ThemeData(
-    appBarTheme: const AppBarTheme(
-        color: Color(0xFFFFFFFF),
-        elevation: 0,
-        iconTheme: IconThemeData(
-            color: Color(0xFF000000)
-        ),
-        titleTextStyle: TextStyle(
-            fontSize: 20,
-            fontWeight: FontWeight.w600,
-            color: Color(0xFF000000))),
+    useMaterial3: true,
     brightness: Brightness.light,
-    primaryColorBrightness: Brightness.dark,
-    primaryColor: Color(0xFFFFFFFF),
-    accentColor: Color(0xFF808184),
-    scaffoldBackgroundColor: Color(0xFFFFFFFF),
-    cardTheme: CardTheme(
-      color: Color(0xFFF5F5F5),
+    primaryColor: const Color(0xFFF0F2F2),
+    canvasColor: const Color(0xFFF0F2F2),
+    scaffoldBackgroundColor: const Color(0xFFF0F2F2),
+    colorScheme: ColorScheme.light(
+        background: const Color(0xFFF0F2F2),
+        primary: Color(0xFF456900),
+        onPrimary: const Color(0xFFFFFFFF),       
+        secondary: Color(0xFF608F06),),
+    popupMenuTheme: const PopupMenuThemeData(
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.all(Radius.circular(12)),
+      ),
+      color: Color(0xFFF5F7F7),
     ),
-    dialogTheme: DialogTheme(
-      backgroundColor: Color(0xFFF9F9F9),
+    appBarTheme: const AppBarTheme(
+        color: Color(0xFFF0F2F2),
+        elevation: 0,
+        iconTheme: IconThemeData(color: Color(0xFF000000)),
+        titleTextStyle: TextStyle(
+            fontSize: 22,
+            fontWeight: FontWeight.w400,
+            color: Color(0xFF000000))),
+    cardTheme: const CardTheme(
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.all(Radius.circular(12)),
+      ),
+      color: Color(0xFFFDFFFF),
     ),
-    bottomAppBarColor: Color(0xFFE6E6E6),
+    dialogTheme: const DialogTheme(
+      backgroundColor: Color(0xFFFDFFFF),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.all(Radius.circular(28)),
+      ),
+    ),
+    floatingActionButtonTheme: FloatingActionButtonThemeData(
+      backgroundColor: Color(0xFF456900),
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.all(Radius.circular(16)),
+      ),
+    ),
+    bottomAppBarColor: const Color(0xFFF0F2F2),
+    bottomNavigationBarTheme: const BottomNavigationBarThemeData(
+      selectedIconTheme: IconThemeData(color: Color(0xFF456900),),
+      selectedLabelStyle: TextStyle(color: Color(0xFF456900),),
+      showSelectedLabels: true,
+      showUnselectedLabels: false,
+      backgroundColor: Color(0xFFF0F2F2),
+    ),
+    navigationBarTheme: NavigationBarThemeData(
+        backgroundColor: const Color(0xFFF0F2F2),
+        indicatorColor: Color(0xFF456900),
+        iconTheme: MaterialStateProperty.all(const IconThemeData(
+          color: Color(0xFF050505),
+        )),
+        labelTextStyle: MaterialStateProperty.all(const TextStyle(
+            color: Color(0xFF050505), fontWeight: FontWeight.w500))),
     bottomSheetTheme:
-        BottomSheetThemeData(modalBackgroundColor: Color(0xFFF5F5F5)));
+    const BottomSheetThemeData(modalBackgroundColor: Color(0xFFF0F2F2)));
 
-//ESCURO
 ThemeData dark = ThemeData(
+    useMaterial3: true,
+    brightness: Brightness.dark,
+    primaryColor: const Color(0xFF202020),
+    scaffoldBackgroundColor: const Color(0xFF202020),
+    canvasColor: const Color(0xFF202020),
+    colorScheme: const ColorScheme.dark(
+        background: Color(0xFF202020),
+        primary: Color(0xFFA6D654),
+        onPrimary: Color(0xFF213600),
+        secondary: Color(0xFFA6D654)),
     appBarTheme: const AppBarTheme(
         color: Color(0xFF202020),
         elevation: 0,
+        iconTheme: IconThemeData(color: Color(0xFFFFFFFF)),
         titleTextStyle: TextStyle(
-            fontSize: 20,
-            fontWeight: FontWeight.w600,
+            fontSize: 22,
+            fontWeight: FontWeight.w400,
             color: Color(0xFFFFFFFF))),
-    brightness: Brightness.dark,
-    primaryColorBrightness: Brightness.light,
-    primaryColor: Color(0xFF202020),
-    accentColor: Color(0xFF484848),
-    scaffoldBackgroundColor: Color(0xFF202020),
-    cardTheme: CardTheme(
+    cardTheme: const CardTheme(
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.all(Radius.circular(12)),
+      ),
       color: Color(0xFF2D2D2D),
     ),
-    dialogTheme: DialogTheme(
-      backgroundColor: Color(0xFF363637),
+    dialogTheme: const DialogTheme(
+      backgroundColor: Color(0xFF2D2D2D),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.all(Radius.circular(28)),
+      ),
     ),
-    bottomAppBarColor: Color(0xFF171717),
+    bottomNavigationBarTheme: const BottomNavigationBarThemeData(
+      selectedIconTheme: IconThemeData(color: Color(0xFFA6D654)),
+      selectedLabelStyle: TextStyle(color: Color(0xFFA6D654)),
+      showSelectedLabels: true,
+      showUnselectedLabels: false,
+      backgroundColor: Color(0xFF202020),
+    ),
+    floatingActionButtonTheme: const FloatingActionButtonThemeData(
+      backgroundColor: Color(0xFFA6D654),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.all(Radius.circular(16)),
+      ),
+    ),
+    navigationBarTheme: NavigationBarThemeData(
+        backgroundColor: const Color(0xFF202020),
+        indicatorColor: const Color(0xFFA6D654),
+        iconTheme: MaterialStateProperty.all(const IconThemeData(
+          color: Color(0xFFEAEAEA),
+        )),
+        labelTextStyle: MaterialStateProperty.all(const TextStyle(
+            color: Color(0xFFEAEAEA), fontWeight: FontWeight.w500))),
+
+    bottomAppBarColor: const Color(0xFF202020),
     bottomSheetTheme:
-        BottomSheetThemeData(modalBackgroundColor: Color(0xFF252525)));
-
-class ThemeNotifier extends ChangeNotifier {
-  final String key = 'valorTema';
-  SharedPreferences prefs;
-  bool _darkTheme;
-
-  bool get darkTheme => _darkTheme;
-
-  ThemeNotifier() {
-    _darkTheme = true;
-    _loadFromPrefs();
-  }
-
-  toggleTheme() {
-    _darkTheme = !_darkTheme;
-    _saveToPrefs();
-    notifyListeners();
-  }
-
-  _initPrefs() async {
-    if (prefs == null) {
-      prefs = await SharedPreferences.getInstance();
-    }
-  }
-
-  _loadFromPrefs() async {
-    await _initPrefs();
-    _darkTheme = prefs.getBool(key) ?? true;
-    notifyListeners();
-  }
-
-  _saveToPrefs() async {
-    await _initPrefs();
-    prefs.setBool(key, _darkTheme);
-  }
-}
+    const BottomSheetThemeData(modalBackgroundColor: Color(0xFF202020)));
