@@ -1,24 +1,31 @@
+import 'package:easy_dynamic_theme/easy_dynamic_theme.dart';
 import 'package:flutter/material.dart';
-import 'package:moodiefschmtz/app.dart';
 import 'package:moodiefschmtz/util/theme.dart';
-import 'package:provider/provider.dart';
+import 'app.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-
-  runApp(ChangeNotifierProvider(
-    create: (_) => ThemeNotifier(),
-    child: Consumer<ThemeNotifier>(
-      builder:(context, ThemeNotifier notifier, child){
-        return MaterialApp(
-          theme: notifier.darkTheme ? dark : light,
-          debugShowCheckedModeBanner: false,
-          home: App(),
-        );
-      },
+  runApp(
+    EasyDynamicThemeWidget(
+      child: const StartAppTheme(),
     ),
-  )
   );
+}
+
+
+class StartAppTheme extends StatelessWidget {
+  const StartAppTheme({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      theme: light,
+      darkTheme: dark,
+      themeMode: EasyDynamicTheme.of(context).themeMode,
+      home:  App(),
+    );
+  }
 }
 
